move rpc errors to meta crates
