+++ crates/client/src/rpc/errors/mod.rs
@@ -1,7 +1,3 @@
-pub mod error_codes;
-mod error_helpers;
-mod errors;
-
-pub use error_codes as codes;
-pub use error_helpers::*;
-pub use errors::*;
+pub use cfx_rpc_utils::error::{
+    error_codes, error_codes as codes, errors::*, jsonrpc_error_helpers::*,
+};
