+++ crates/rpc/rpc-eth-types/Cargo.toml
@@ -29,4 +29,3 @@ cfx-vm-types = { workspace = true }
 cfx-rpc-cfx-types = { workspace = true }
 error-chain = { workspace = true }
 cfx-parity-trace-types = { workspace = true }
-
