+++ crates/rpc/rpc-utils/src/error/mod.rs
@@ -1,0 +1,4 @@
+pub mod error_codes;
+pub mod errors;
+pub mod jsonrpc_error_helpers;
+pub mod jsonrpsee_error_helpers;
